requirements.txt
fastapi
ta-lib-bin
httpx